Use an integer to store time and date values internally.  The use of a double was causing some inaccuracies to appear when filling columns of times
svn path=/trunk/koffice/; revision=515846

+++ kspread/valueconverter.cc
@@ -315,7 +315,7 @@ Value ValueConverter::asTime (const Value &value) const
       val.setValue (QTime::currentTime());
     break;
     case Value::Integer:
-      val.setValue (value.asFloat());
+      val.setValue (value.asInteger());
       val.setFormat (Value::fmt_Time);
     break;
     case Value::Float:
